Introduce a search feature, replacing tag groups
Summary:
When you load, say, the histograms dashboard, you now see this:

![New view of the histogram dashboard.](https://user-images.githubusercontent.com/4317806/27709831-71f3a24e-5cd2-11e7-8d66-410f6e6bfefa.png)

By introducing a search bar in the main content pane, we get the
following features for free.

First, of course, we have search itself. This is superior to tag groups
because updating the search query does not lose the state of all
categories (both the category panes' "am I expanded?" state and the
individual visualization's properties, like the current plot window).
This also means that you can see your search results update as you type.

Second, we set the default search query to `.*`. This has the effect
that the initial screen of a dashboard provides useful data instead of
just collapsed categories loading a dashboard shows you some useful data
instead of nothing. It also suggests to the user that the search field
supports regular expressions, which is a nice bonus.

Test Plan:
For the unit tests: use the processed described in #146.
![Proof of passing tests](https://user-images.githubusercontent.com/4317806/27711999-e62c4272-5cd9-11e7-89c8-16cf0762c797.png)

wchargin-branch: search